Fix module 9 pdf, clean code chunks from modules
There is no reason to use comment = '', prompt = TRUE in the modules, so I
removed them. The computations under cache = TRUE were fast enough, so the best
was just to remove them to avoid issues between rendering html/pdfs. As for the
actual issue with module 9, it had to do with centering the images which is
valid in the html format but broke in the pdf one.